+++ sdk/core/azure-core/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Breaking Changes
 
 - Removed `Context::GetApplicationContext()` in favor of a new static data member `Context::ApplicationContext`.
+- Renames `IsDownloadViaStream()` to `IsNotBufferDownload()`.
 
 ### Bug Fixes